Move `separators.hpp` to `layout.hpp`
